Added client-side prediction for Battery and PowerCell
