+++ Content.Server/Holosign/HolosignSystem.cs
@@ -1,6 +1,6 @@
 using Content.Shared.Examine;
 using Content.Shared.Coordinates.Helpers;
-using Content.Server.Power.Components;
+using Content.Shared.Power.Components;
 using Content.Server.PowerCell;
 using Content.Shared.Interaction;
 using Content.Shared.Storage;
