+++ Content.Server/Light/EntitySystems/EmergencyLightSystem.cs
@@ -8,6 +8,7 @@ using Content.Server.Station.Systems;
 using Content.Shared.Examine;
 using Content.Shared.Light;
 using Content.Shared.Light.Components;
+using Content.Shared.Power.Components;
 using Robust.Server.GameObjects;
 using Color = Robust.Shared.Maths.Color;
