+++ Content.Server/Ninja/Systems/BatteryDrainerSystem.cs
@@ -6,7 +6,7 @@ using Content.Shared.Interaction;
 using Content.Shared.Ninja.Components;
 using Content.Shared.Ninja.Systems;
 using Content.Shared.Popups;
-using Robust.Shared.Audio;
+using Content.Shared.Power.Components;
 using Robust.Shared.Audio.Systems;
 
 namespace Content.Server.Ninja.Systems;
