+++ Content.Server/Ninja/Systems/NinjaSuitSystem.cs
@@ -1,12 +1,12 @@
 using Content.Server.Emp;
 using Content.Server.Ninja.Events;
-using Content.Server.Power.Components;
 using Content.Server.PowerCell;
 using Content.Shared.Clothing.EntitySystems;
 using Content.Shared.Hands.EntitySystems;
 using Content.Shared.Ninja.Components;
 using Content.Shared.Ninja.Systems;
 using Content.Shared.Popups;
+using Content.Shared.Power.Components;
 using Content.Shared.PowerCell.Components;
 using Robust.Shared.Containers;
