+++ Content.Server/Ninja/Systems/SpaceNinjaSystem.cs
@@ -21,6 +21,7 @@ using Robust.Shared.Player;
 using System.Diagnostics.CodeAnalysis;
 using Content.Server.Objectives.Components;
 using Robust.Shared.Audio.Systems;
+using Content.Shared.Power.Components;
 
 namespace Content.Server.Ninja.Systems;
