--- Content.Server/Power/Components/ExaminableBatteryComponent.cs
@@ -1,6 +1,0 @@
-namespace Content.Server.Power.Components
-{
-    [RegisterComponent]
-    public sealed partial class ExaminableBatteryComponent : Component
-    {}
-}
