+++ Content.Server/Power/SetBatteryPercentCommand.cs
@@ -1,7 +1,7 @@
 using Content.Server.Administration;
-using Content.Server.Power.Components;
 using Content.Server.Power.EntitySystems;
 using Content.Shared.Administration;
+using Content.Shared.Power.Components;
 using Robust.Shared.Console;
 
 namespace Content.Server.Power
