+++ Content.Server/SensorMonitoring/BatterySensorSystem.cs
@@ -2,6 +2,7 @@
 using Content.Server.DeviceNetwork.Systems;
 using Content.Server.Power.Components;
 using Content.Shared.DeviceNetwork;
+using Content.Shared.Power.Components;
 
 namespace Content.Server.SensorMonitoring;
