+++ Content.Shared/Power/ExaminableBatteryComponent.cs
@@ -1,0 +1,8 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Power.Components;
+
+[RegisterComponent, NetworkedComponent]
+public sealed partial class ExaminableBatteryComponent : Component
+{
+}
