+++ Content.Shared/PowerCell/Components/PowerCellSlotComponent.cs
@@ -1,8 +1,9 @@
 using Content.Shared.Containers.ItemSlots;
+using Robust.Shared.GameStates;
 
 namespace Content.Shared.PowerCell.Components;
 
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
 public sealed partial class PowerCellSlotComponent : Component
 {
     /// <summary>
@@ -12,14 +13,14 @@ public sealed partial class PowerCellSlotComponent : Component
     /// Given that <see cref="PowerCellSystem"/> needs to verify that a given cell has the correct cell-size before
     /// inserting anyways, there is no need to specify a separate entity whitelist. In this slot's yaml definition.
     /// </remarks>
-    [DataField("cellSlotId", required: true)]
+    [DataField(required: true), AutoNetworkedField]
     public string CellSlotId = string.Empty;
 
     /// <summary>
     /// Can this entity be inserted directly into a charging station? If false, you need to manually remove the power
     /// cell and recharge it separately.
     /// </summary>
-    [DataField("fitsInCharger")]
+    [DataField]
     public bool FitsInCharger = true;
 
 }
@@ -27,12 +28,7 @@ public sealed partial class PowerCellSlotComponent : Component
 /// <summary>
 ///     Raised directed at an entity with a power cell slot when the power cell inside has its charge updated or is ejected/inserted.
 /// </summary>
-public sealed class PowerCellChangedEvent : EntityEventArgs
+public sealed class PowerCellChangedEvent(bool ejected) : EntityEventArgs
 {
-    public readonly bool Ejected;
-
-    public PowerCellChangedEvent(bool ejected)
-    {
-        Ejected = ejected;
-    }
+    public readonly bool Ejected = ejected;
 }
